fix cache and artifacts
