+++ packages/hardhat-zksync-solc/src/constants.ts
@@ -1,0 +1,4 @@
+export const ARTIFACT_FORMAT_VERSION = "hh-zksolc-artifact-1";
+export const DEBUG_NAMESPACE = "hardhat:plugin:zksolc";
+export const CACHE_FORMAT_VERSION = "hh-zksolc-cache-1";
+export const ZKSOLC_FILES_CACHE_FILENAME = "zksolc-files-cache.json";
